chore: move graphql under store and fix imports
