+++ codegen.yml
@@ -1,6 +1,6 @@
 overwrite: true
 schema: "https://raw.githubusercontent.com/GaloyMoney/galoy/main/src/graphql/main/schema.graphql"
-documents: "src/graphql/*/*.ts"
+documents: "src/store/graphql/*/*.ts"
 generates:
   src/graphql/types.ts:
     config:
