+++ src/store/graphql/mutations/device-notification-token-create.ts
