+++ src/store/graphql/mutations/ln-no-amount-invoice-create.ts
