+++ src/store/graphql/mutations/user-quiz-question-update-completed.ts
