+++ src/store/graphql/queries/btc-price-list.ts
