+++ src/store/graphql/queries/currency-list.ts
