+++ src/store/graphql/queries/main.ts
