+++ src/store/graphql/queries/quiz-questions.ts
