+++ src/store/graphql/queries/user-default-wallet-id.ts
