+++ src/store/graphql/types.ts
